Move classes to avoid undesired dependencies.
