+++ org.jgrapes.webconsole.base/src/org/jgrapes/webconsole/base/events/UserAuthenticated.java
@@ -16,7 +16,7 @@
  * with this program; if not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.jgrapes.webconsole.rbac;
+package org.jgrapes.webconsole.base.events;
 
 import java.util.ArrayList;
 import java.util.List;
